feat: 스크랩 취소 데이터 전송 초기 세팅 및 컴포넌트 구조 분리
