+++ src/app/albatalk/[id]/components/DetailContainer.tsx
@@ -7,7 +7,7 @@ import Modal from '@/components/modal/Modal';
 import DetailSkeleton from './DetailSkeleton';
 import { useLikePosts } from '@/hooks/mutation/useLikePosts';
 import LikeButton from './LikeButton';
-import { useState } from 'react';
+import { useImgError } from '@/hooks/common/useImgError';
 
 export default function DetailContainer({
   userId,
@@ -25,17 +25,13 @@ export default function DetailContainer({
   isShowComments,
   totalCommentCount,
 }: PostDetailProps) {
-  const [profileImg, setProfileImg] = useState<Record<string, string>>({});
-
   const { id: postId, writer } = post ?? {};
 
   const { mutate: toggleLikePost, isPending } = useLikePosts();
 
-  const defaultProfileImg = '/images/defaultProfile.svg';
-  const handleProfileImgError = (src: string) => {
-    setProfileImg((prev) => ({ ...prev, [src]: defaultProfileImg }));
-  };
-
+  const { img, defaultImg, handleImgError } = useImgError(
+    '/images/defaultProfile.svg',
+  );
   return (
     <>
       {showModal && modalType === 'deletePost' ? (
@@ -73,15 +69,15 @@ export default function DetailContainer({
             <div className='flex items-center'>
               <Image
                 src={
-                  profileImg[String(writer?.imageUrl)] ||
+                  img[String(writer?.imageUrl)] ||
                   writer?.imageUrl ||
-                  defaultProfileImg
+                  defaultImg
                 }
                 alt='기본프로필'
                 width={26}
                 height={26}
                 className='rounded-[50%] object-cover border border-gray500 min-h-[26px]'
-                onError={() => handleProfileImgError(String(writer?.imageUrl))}
+                onError={() => handleImgError(String(writer?.imageUrl))}
               />
               <p className='ml-[10px]'>{writer?.nickname}</p>
             </div>
